Update ENSEMBLE.R
If both methods (BR and EM) are present when using the lad function, it will report an error:

Error in switch(method, BR = lad.fit.BR(x, y, tol), EM = lad.fit.EM(x,  : 
  EXPR must be a length 1 vector

If we only use one of the methods, it will work fine. I tested both methods using the bulk RNA-seq in the tutorial, and it seems that the results are similar. So, I think going with the default will solve the problem.